+++ src/test/php/ValueEqualsTest.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 /**
  * This file is part of stubbles.
  *
@@ -34,7 +35,7 @@ class EqualsTest extends \PHPUnit_Framework_TestCase
     /**
      * @return  array
      */
-    public function tuplesEvaluatingToTrue()
+    public function tuplesEvaluatingToTrue(): array
     {
         return [[true, true],
                 [false, false],
@@ -58,7 +59,7 @@ class EqualsTest extends \PHPUnit_Framework_TestCase
     /**
      * @return  array
      */
-    public function tuplesEvaluatingToFalse()
+    public function tuplesEvaluatingToFalse(): array
     {
         return [[true, false],
                 [false, true],